requirements.txt
python-dotenv==0.19.2
vk_api==11.9.7
python-telegram-bot==13.8.1
